extracting what can be
